requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
langchain==0.0.255
flask
pinecone-client
pypdf
python_dotenv==1.0.0
-e .